requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
numpy
requests==2.24.0
gunicorn==20.1.0
Pillow==8.3.2
tensorflow==2.10.0
